requirements.txt
requests==2.22.0
djangorestframework==3.10.0
Django==2.2
numpy==1.19.5
pandas==1.3.3
python-dotenv==0.19.2
pymongo==3.12.0
python-dotenv==0.19.2
django-cors-headers==3.4.0
mssql-django==1.0
django-mssql-backend==2.8.1